move name_giver.py from /scripts to /src/jetson folder
